refactor(acp): split ZephAcpAgentState god-object into focused modules
ZephAcpAgentState accumulated three impl blocks (~2955 lines) spanning
six unrelated responsibilities: builder wiring, idle-session reaping,
LSP diagnostics, session lifecycle, slash-command/model dispatch, and
MCP extension handling. This increased merge-conflict surface on every
ACP feature PR and made it hard to reason about which methods were
safe to test in isolation.

Extract each cluster into a sibling file under crates/zeph-acp/src/agent/
following the existing providers.rs/usage.rs precedent (file-local impl
blocks on ZephAcpAgentState, pub(crate) bumps only where cross-file
calls require it): builder.rs, reaper.rs, lsp_events.rs, mcp_ext.rs,
model.rs, slash.rs, turn.rs, session.rs. mod.rs shrinks from 4424 to
1408 lines and keeps the struct definition, type aliases, and thin
protocol dispatch methods.

Mechanical extraction only: all 25 fields stay on the coordinator
struct, no behavior change, no signature changes beyond the necessary
private-to-pub(crate) visibility bumps.
